requirements.txt
streamlit>=1.28.0
langchain==1.0.8
langchain-community==0.4.1
langchain-google-genai==3.1.0
langchain-core==1.1.0
crewai>=0.28.0
litellm>=1.80.0
neo4j>=5.14.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
pyvis>=0.3.2
networkx>=3.2.0
pydantic>=2.0.0,<3.0.0
google-generativeai>=0.3.0

